refactor scanner and scanner utils
